Sync: 115. Distinct Subsequences (Hard) - Leet2Git

+++ 115_Distinct_Subsequences.cpp
@@ -21,12 +21,14 @@ public:
         int m = s.size();
         int n = t.size();
 
-        vector<int>  dp(n + 1, 0);
+        vector<long long>  dp(n + 1, 0);
     
         dp[n] = 1;
         
         for(int i = m - 1; i >= 0; i--) {
-            vector<int> curr = dp;
+
+            vector<long long> curr = dp;
+
             for(int j = n - 1; j >= 0; j--) {
                     if(s[i] == t[j]) {
                     curr[j] = dp[j + 1] + dp[j];
@@ -39,6 +41,6 @@ public:
             dp = curr;
         }
 
-        return dp[0];
+        return (int)dp[0];
     }
 };
